chapters/2023/Structure de graphe au service de l'analyse de pipeline CI-CD/assets/codes/requirements.txt
pyyaml
graphviz
GitPython
rich